refactor(menu): split executor.go into cohesive files (pure code movement)
- executor.go keeps types, constructor, accessors, small session helpers
- executor_run.go: Run loop + executeCommandAction dispatch
- executor_display.go: template tokens, displayFile, displayPrompt, file includes
- executor_runnables.go: runnable registries + core RUN: command handlers
- executor_session.go: session InputHandler, cursor, holdScreen, line readers
- oneliner mutex/consts/types moved to executor_oneliners.go
- lastCallerATTokenRegex moved to executor_lastcallers.go
- lightbar consts/LightbarOption/ansi color maps moved to executor_lightbar.go

No behavior change; executor_run.go stays at 958 lines because Run() is a
single ~830-line function.
